gjs 1.54: compatibility ByteArray.toString
gjs 1.54 deprecates the usage of .toString() on returned ByteArrays.
Since 1.54 returns Uint8Array instead of the custom ByteArray and
deprecats uint8_array.toString() as UTF-8 decoding function as the
standard expects uint8_array to return a string with the decimal
digits of each byte joined with commas.

This commit adds a compatibility wrapper to work correctly with both gjs
1.52 and 1.54 by checking the instance type of the array to call the
correct function to return a UTF-8 decoded string.

https://gitlab.gnome.org/GNOME/gjs/blob/master/NEWS#L145
https://gitlab.gnome.org/GNOME/gjs/blob/master/NEWS#L4