gjs 1.54: don't define properties as let or const
This fixes the following warnings:
`Some code accessed the property 'x' on the module 'utilities'.
That property was defined with 'let' or 'const' inside the module.
This was previously supported, but is not correct according to the ES6 standard.
Any symbols to be exported from a module must be defined with 'var'.
The property access will work as previously for the for the time being, but
please fix your code anyway.`